M2a: snapshot wire format — proto3 schema, codec, content digests
The snapshot is a contract between a Python control plane that compiles it and
a Go data plane that serves from it, so it needs one definition generated for
both rather than two hand-maintained ones that drift.

  - proto/gateway/v1/snapshot.proto is the schema. Layers are transferable
    individually, which is the point of the layering: a budget edit ships one
    small tenant layer rather than the whole catalog.
  - internal/wire converts between the schema and the domain types. Conversion
    goes through core's Spec types, never into or out of a built layer, so the
    direction of travel is explicit: bytes -> spec -> validated layer.
  - Content digests. A layer is hashed with its own digest field cleared, over
    deterministically marshalled bytes, so a worker can skip a re-fetch it
    already holds and two workers claiming a version can be proven to agree.
    This is a corruption check, not an authenticity one — anyone who can
    rewrite the layer can rewrite the digest. Authenticity is the transport's.

Validation stays in core. The wire layer shapes data and core decides whether
it is coherent, so there is one definition of a valid snapshot rather than a
second, weaker one here.

Enums convert through explicit tables rather than numeric casts. The numbering
agrees today; a cast would make that coincidence load-bearing, so reordering a
domain constant would silently change what a stored snapshot means. An unknown
enum value fails the layer instead of defaulting to the most permissive one.

Two bugs the round-trip tests caught, fixed at the source:

  - core.Wrap returned *Error, so `return core.Wrap(...)` into an `error`
    result produced a non-nil interface holding a nil pointer: an error that is
    not nil and has no message. It now returns `error`, removing the trap for
    every caller rather than patching the one site that hit it.
  - Deployment.Weight was uint8 and HeadroomBasisPoints uint16, both narrower
    than the wire. An out-of-range value truncated into a plausible one — 300
    became 44 — instead of failing core's range check. Both widened to match.

Also: proto3 rather than editions, since protoc 25.x treats editions as
experimental. Generated code is committed so the build needs no toolchain, and
CI regenerates it to prove it still matches the schema.

+++ dataplane/internal/core/catalog.go
@@ -74,10 +74,13 @@ type Deployment struct {
 	// the control plane's SecretsPort and cache the result with a TTL.
 	CredentialRef string
 
-	// Weight is the share of traffic this deployment may take, 0-100. Zero means
+	// Weight is the share of traffic this deployment may take, 0-100. It is
+	// uint32 to match the wire format: a narrower domain type would truncate an
+	// out-of-range value into a plausible one instead of failing validation.
+	// Zero means
 	// registered but not serving: the state a new fine-tuned adapter sits in
 	// while it takes shadow traffic, before the canary steps begin.
-	Weight uint8
+	Weight uint32
 
 	Cost         Cost
 	Capabilities []Capability
